+++ src/api/api.ts
@@ -829,8 +829,8 @@ export default class MastoApi {
 
                 // breakIf() must be called before we check the length of rows!  // TODO: still necessary?
                 const shouldStop = breakIf ? (await breakIf(page, newRows)) : false;
-                let resultsMsg = `got page ${++pageNumber} with ${page.length} objs ${waitTime.ageString()}`;
-                resultsMsg += `, ${newRows.length} objs so far`;
+                const resultsMsg = `got page ${++pageNumber} with ${page.length} objs ${waitTime.ageString()}` +
+                                   `, ${newRows.length} objs so far`;
 
                 if (newRows.length >= maxRecords || page.length == 0 || shouldStop) {
                     const msg = `Fetch finished (${resultsMsg}, shouldStop=${shouldStop}, maxRecords=${maxRecords})`;
@@ -932,7 +932,7 @@ export default class MastoApi {
             // If endpoint has unique IDs use both cached and new rows (it's deduped in buildFromApiObjects())
             // newRows are in front so they will survive truncation (if it happens)
             newRows = UNIQUE_ID_PROPERTIES[cacheKey] ? [...newRows, ...cachedRows] : newRows;
-            const objs = this.buildFromApiObjects<T>(cacheKey, newRows as T[], logger);
+            const objs = await this.buildFromApiObjects<T>(cacheKey, newRows as T[], logger);
 
             // If we have a maxCacheRecords limit, truncate the new rows to that limit
             if (maxCacheRecords && objs.length > maxCacheRecords) {
@@ -1127,7 +1127,11 @@ export default class MastoApi {
      * @param {Logger} logger - Logger instance.
      * @returns {ApiObj[]} Array of constructed objects.
      */
-    private buildFromApiObjects<T extends ApiObj>(key: CacheKey, objects: T[], logger: Logger): ResponseRow<T>[] {
+    private async buildFromApiObjects<T extends ApiObj>(
+        key: CacheKey,
+        objects: T[],
+        logger: Logger
+    ): Promise<ResponseRow<T>[]> {
         let newObjects: ResponseRow<T>[];
         const nullObjs = objects.filter(isNil);
 
@@ -1136,9 +1140,7 @@ export default class MastoApi {
         }
 
         if (STORAGE_KEYS_WITH_TOOTS.includes(key)) {
-            const toots = objects.map(obj => Toot.build(obj as TootLike));
-            // Toots get special handling for deduplication
-            return Toot.dedupeToots(toots, logger.tempLogger(`buildFromApiObjects`)) as ResponseRow<T>[];
+            return await Toot.buildToots(objects as TootLike[], key) as ResponseRow<T>[];  // Toots have special dedupe handling
         } else if (STORAGE_KEYS_WITH_ACCOUNTS.includes(key)) {
             newObjects = objects.map(obj => Account.build(obj as AccountLike)) as ResponseRow<T>[];
         } else {