+++ src/api/objects/toot.ts
@@ -924,18 +924,18 @@ export default class Toot implements TootObj {
         const userData = await MastoApi.instance.getUserData();
         const trendingTags = (await MastodonServer.fediverseTrendingTags()).topObjs();
         const trendingLinks = isDeepInspect ? (await MastodonServer.fediverseTrendingLinks()) : []; // Skip trending links
-        let completeToots: TootLike[] = [];
-        let tootsToComplete = toots;
+        let completedToots: TootLike[] = [];
+        let incompleteToots = toots;
 
         // If isDeepInspect separate toots that need completing bc it's slow to rely on isComplete() + batching
         if (isDeepInspect) {
-            [completeToots, tootsToComplete] = (split(toots, (t) => t instanceof Toot && t.isComplete()));
+            [completedToots, incompleteToots] = (split(toots, (t) => t instanceof Toot && t.isComplete()));
         }
 
         const newCompleteToots: Toot[] = await batchMap(
-            tootsToComplete,
+            incompleteToots,
             async (tootLike: TootLike) => {
-                const toot = (tootLike instanceof Toot ? tootLike : Toot.build(tootLike));
+                const toot = (tootLike instanceof Toot) ? tootLike : Toot.build(tootLike);
                 toot.completeProperties(userData, trendingLinks, trendingTags, source);
                 return toot as Toot;
             },
@@ -947,8 +947,8 @@ export default class Toot implements TootObj {
         );
 
         const msg = `${toots.length} toots ${ageString(startedAt)}`;
-        logger.debug(`${msg} (${newCompleteToots.length} completed, ${completeToots.length} skipped)`);
-        return newCompleteToots.concat(completeToots as Toot[]);
+        logger.debug(`${msg} (${newCompleteToots.length} completed, ${completedToots.length} skipped)`);
+        return newCompleteToots.concat(completedToots as Toot[]);
     }
 
     /**
@@ -981,9 +981,10 @@ export default class Toot implements TootObj {
             let reblogsBy = this.uniqFlatMap<Account>(uriToots, "reblogsBy", (account) => account.webfingerURI);
             reblogsBy = sortObjsByProps(reblogsBy, ["displayName"], true, true);
             // Collate accounts - reblogs and realToot accounts
-            const allAccounts = uriToots.flatMap(t => [t.account].concat(t.reblog ? [t.reblog.account] : []));
+            const allAccounts = uriToots.flatMap(toot => toot.accounts);
             // Helper method to collate the isFollowed property for the accounts
             const isFollowed = (uri: string) => allAccounts.some((a) => a.isFollowed && (a.webfingerURI == uri));
+            const isSuspended = (uri: string) => allAccounts.some((a) => a.suspended && (a.webfingerURI == uri));
 
             // Counts may increase over time w/repeated fetches so we collate the max
             const propsThatChange = PROPS_THAT_CHANGE.reduce((propValues, propName) => {
@@ -1011,12 +1012,15 @@ export default class Toot implements TootObj {
                 toot.realToot.bookmarked = uriToots.some(toot => toot.realToot.bookmarked);
                 toot.realToot.favourited = uriToots.some(toot => toot.realToot.favourited);
                 toot.realToot.reblogged = uriToots.some(toot => toot.realToot.reblogged);
-                toot.account.isFollowed ||= isFollowed(toot.account.webfingerURI);
                 toot.muted = uriToots.some(toot => toot.muted || toot.realToot.muted);  // Liberally set muted on retoots and real toots
+
+                toot.accounts.forEach((account) => {
+                    account.isFollowed ||= isFollowed(account.webfingerURI);
+                    account.suspended ||= isSuspended(account.webfingerURI);
+                });
 
                 // Reblog props
                 if (toot.reblog) {
-                    toot.reblog.account.isFollowed ||= isFollowed(toot.reblog.account.webfingerURI);
                     toot.reblog.completedAt ??= lastCompleted?.realToot.completedAt;
                     toot.reblog.filtered = uniqFiltered;
                     toot.reblog.reblogsBy = reblogsBy;